fixes file naming for non-jsx files
