+++ src/to-be-deleted/configuration.ts
